Explicitly cast upper_left and lower_right to ints

+++ skimage/draw/draw.py
@@ -58,8 +58,8 @@ def ellipse(cy, cx, yradius, xradius, shape=None):
 
     # The upper_left and lower_right corners of the
     # smallest rectangle containing the ellipse.
-    upper_left = np.ceil(center - radiuses)
-    lower_right = np.floor(center + radiuses)
+    upper_left = np.ceil(center - radiuses).astype(int)
+    lower_right = np.floor(center + radiuses).astype(int)
 
     if shape is not None:
         # Constrain upper_left and lower_right by shape boundary.
